Moving out the `components` variable to its own file that we can generate using code
