+++ package.json
@@ -7,6 +7,7 @@
   "exports": "./src/index.tsx",
   "scripts": {
     "postinstall": "husky install",
+    "gen": "node src/layout/components.gen.mjs",
     "start": "cross-env NODE_ENV=development webpack-dev-server --config webpack.config.development.js --mode development --progress",
     "start:silent": "cross-env NOTIFY_ON_ERRORS=false yarn start",
     "build": "cross-env NODE_ENV=production webpack --config webpack.config.production.js --progress",
