+++ src/layout/components.gen.mjs
@@ -1,0 +1,34 @@
+import fs from 'node:fs';
+import path from 'node:path';
+
+const components = {};
+
+fs.readdirSync('src/layout').forEach((item) => {
+  try {
+    const stat = fs.statSync(path.join('src/layout', item));
+    if (stat.isDirectory()) {
+      if (item === 'Address') {
+        // Address is a special case, because we once named it 'AddressComponent', without any over the other components
+        // having that suffix. We need to keep this for backwards compatibility, but our folder structure uses the name
+        // without the suffix.
+        components[item] = 'AddressComponent';
+        return;
+      }
+
+      components[item] = item;
+    }
+  } catch (e) {}
+});
+
+const lines = [
+  '// This file is generated by running src/layout/components.gen.ts',
+  '',
+  ...Object.keys(components).map((key) => `import { ${key} } from 'src/layout/${key}/index';`),
+  '',
+  `export const components = {`,
+  ...Object.keys(components).map((key) => `  ${components[key]}: new ${key}(),`),
+  `};`,
+  '',
+];
+
+fs.writeFileSync('src/layout/components.ts', lines.join('\n'));
